refactor(ic-stable-lara): optimize counterpart lookup with fast paths
Split counterpart resolution into fast and verified variants. The fast
path
stops after locating the counterpart, adding early singleton and
self-loop shortcuts and new rank/select primitives on LabeledLaraGraph.

+++ crates/graph/src/bench/mate_adoption_gate.rs
@@ -4577,7 +4577,7 @@ mod fixture_evidence_tests {
                     owner: identity.owner,
                     target: identity.target,
                     orientation: identity.orientation,
-                    slot: identity.slot as u32,
+                    slot: identity.slot,
                 },
             )
             .collect::<Vec<_>>();
@@ -4597,7 +4597,7 @@ mod fixture_evidence_tests {
                     owner: identity.owner,
                     target: identity.target,
                     orientation: identity.orientation,
-                    slot: identity.slot as u32,
+                    slot: identity.slot,
                 },
             )
             .collect::<Vec<_>>();
